Support nested before(:all) hooks
Because selective splits by example before all hooks must be run as
before each hooks. We got this working recently but it was not working
when defined in a nested context. This commit fixes that by explicitly
changing before all hooks to before each hooks.

This is a behavior we may enable with an explicit flag or configuration
option in the future. For now it is enabled by default.